fixup! Extend try...except translation support
